+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - FEATURE: add `n_retry` test set property. When set, each case will be tried that at most that many times until first success (if any).
 - FEATURE: add `hooks.pre_source` and `hooks.post_source` fields to collected cases. They can be either `'once'` or `'case'` and allow a more granular control over case execution.
+- FEATURE: `finally()` now can be called several times inside a single function with callbacks executed in order of how they were registered.
 
 
 # Version 0.14.0